from float to double

+++ function.h
@@ -4,7 +4,7 @@
 class Function {
   private:
   int _nPoints;
-  float* _f;
+  double* _f;
 
   public:
   Function();                               //default constructor
@@ -15,10 +15,10 @@ class Function {
   Function& operator=(Function &&src);      //move assignment
   ~Function();                              //destructor
 
-  float operator[](int idx) const;         //idx element of _f
-  float& operator[](int idx);              //idx element of _f
+  double operator[](int idx) const;         //idx element of _f
+  double& operator[](int idx);              //idx element of _f
 
-  float* f() const;                        //get _f
+  double* f() const;                        //get _f
   void del();                               //set _f to nullptr
 };
